+++ front-page.php
@@ -2,7 +2,7 @@
 
 $context = \Timber\Timber::get_context();
 
-$context ['post'] = new \App\Post();
+$context ['post'] = new \Timber\Post();
 
 \Timber\Timber::render([
     'views/front-page.twig',