+++ tailwind.config.js
@@ -1,6 +1,23 @@
 module.exports = {
   theme: {
-    extend: {}
+    extend: {},
+    colors: {
+      black: {
+        default: 'hsl(0, 0%, 0%)',
+        venta: 'hsl(0, 0%, 0%)',
+        raw: 'hsl(0, 0%, 13%)',
+        bare: 'hsl(0, 0%, 20%)',
+      },
+      white: {
+        default: 'hsl(0, 0%, 100%)',
+        pure: 'hsl(0, 0%, 100%)',
+        broken: 'hsl(0, 0%, 96%)',
+      },
+      primary: {
+        default: ''
+      },
+      secondary: {}
+    }
   },
   variants: {},
   plugins: []