+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
-## 0.0.1
+## prerelease
 
+- Added elementor theme locations
+- Added basic scripts
+- Added stylesheets
 - We created an initial theme